fix(manifest-ui): fix circular dependency in blogging components
- Created registry/blogging/types.ts with Post interface
- Updated demo/data.ts to import from ../types instead of ../post-card
- Updated post-card.tsx to import and re-export Post from ./types
- Updated post-list.tsx and post-detail.tsx to import Post from ./types
- Added types.ts and demo/data.ts to registry.json files array
- Bumped versions: post-card 2.0.5, post-list 2.0.6, post-detail 2.0.6

This fixes the circular dependency: post-card -> demo/data -> post-card
